pyproject.toml
[tool.poetry]
name = "curtainbe"
version = "0.1.0"
description = ""
authors = ["Toan Phung <toan.phung@proteo.info>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">3.9,<3.13"
django = "^4.2.7"
djangorestframework = "^3.14.0"
django-filter = "^23.3"
djangorestframework-simplejwt = "^5.3.0"
drf-url-filters = "^0.5.1"
django-cors-headers = "^4.3.0"
pandas = "^2.1.2"
uniprotparser = "^1.1.0"
drf-flex-fields = "^1.0.2"
psycopg2 = "^2.9.9"
gunicorn = "^21.2.0"
dj-rest-auth = "^5.0.1"
django-dbbackup = "^4.0.2"
django-rest-auth = "^0.9.5"
django-request = "^1.6.3"
statsmodels = "^0.14.0"
channels = "^4.0.0"
uvicorn = {extras = ["standard"], version = "^0.24.0.post1"}
channels-redis = {extras = ["cryptography"], version = "^4.1.0"}
boto3 = "^1.28.82"
django-rq = "^2.8.1"
djangorestframework-api-key = "^3.0.0"
google-cloud-storage = "^2.13.0"
django-storages = {extras = ["s3"], version = "^1.14.2"}
urllib3 = ">=1.25.4,<1.27"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
